+++ src/ShapeCrawler/ShapeCollection/IShapes.cs
@@ -41,6 +41,6 @@ public interface IShapes : IReadOnlyList<IShape>
     ///     Gets shape by specified type.
     /// </summary>
     /// <typeparam name="T">Shape type.</typeparam>
-    IShape Last<T>()
+    T Last<T>()
         where T : IShape;
 }