+++ src/ShapeCrawler/ShapeCollection/Shapes.cs
@@ -38,7 +38,7 @@ internal sealed class Shapes : IShapes
 
     public IShape GetByName(string name) => this.ShapesCore().First(shape => shape.Name == name);
 
-    public IShape Last<T>() where T : IShape => this.ShapesCore().Last(shape => shape is T);
+    public T Last<T>() where T : IShape => (T)this.ShapesCore().Last(shape => shape is T);
 
     public IEnumerator<IShape> GetEnumerator() => this.ShapesCore().GetEnumerator();