+++ src/ShapeCrawler/ShapeCollection/SlideShapes.cs
@@ -497,12 +497,28 @@ internal sealed class SlideShapes : ISlideShapes
     
     public IShape GetByName(string name) => this.shapes.GetByName(name);
 
-    public IShape Last<T>() where T : IShape => this.shapes.Last<T>();
+    public T Last<T>() where T : IShape => this.shapes.Last<T>();
 
     public IEnumerator<IShape> GetEnumerator() => this.shapes.GetEnumerator();
 
     IEnumerator IEnumerable.GetEnumerator() => this.GetEnumerator();
 
+    private static string Mime(Stream imageStream)
+    {
+        imageStream.Seek(0, SeekOrigin.Begin);
+        using var codec = SKCodec.Create(imageStream);
+        var mime = codec.EncodedFormat switch
+        {
+            SKEncodedImageFormat.Jpeg => "image/jpeg",
+            SKEncodedImageFormat.Png => "image/png",
+            SKEncodedImageFormat.Gif => "image/gif",
+            SKEncodedImageFormat.Bmp => "image/bmp",
+            _ => "image/png"
+        };
+        
+        return mime;
+    }
+    
     private static SizeF GetSvgPixelSize(SvgDocument image)
     {
         // Default base size come from viewbox if specified, else use the raw
@@ -585,7 +601,10 @@ internal sealed class SlideShapes : ISlideShapes
     private P.Picture CreatePPicture(Stream imageStream, string shapeName)
     {
         var imgPartRId = this.sdkSlidePart.NextRelationshipId();
-        var imagePart = this.sdkSlidePart.AddNewPart<ImagePart>("image/png", imgPartRId);
+        var mStream = new MemoryStream();
+        imageStream.CopyTo(mStream);
+        var mime = Mime(mStream);
+        var imagePart = this.sdkSlidePart.AddNewPart<ImagePart>(mime, imgPartRId);
         imageStream.Position = 0;
         imagePart.FeedData(imageStream);
 
@@ -627,7 +646,7 @@ internal sealed class SlideShapes : ISlideShapes
 
         return pPicture;
     }
-
+    
     private void AddPictureSvg(SvgDocument image, Stream svgStream)
     {
         // Determine intrinsic size in