+++ test/ShapeCrawler.Tests.Unit/ShapeCollectionTests.cs
@@ -438,7 +438,7 @@ public class ShapeCollectionTests : SCTest
         // Arrange
         var pres = new Presentation();
         var shapes = pres.Slides[0].Shapes;
-        var image = TestHelper.GetStream("test-image-large.png");
+        var image = TestHelper.GetStream("png image-large.png");
         image.Position = 0;
 
         // Act
@@ -508,7 +508,7 @@ public class ShapeCollectionTests : SCTest
         // Arrange
         var pres = new Presentation();
         var shapes = pres.Slides[0].Shapes;
-        var image = TestHelper.GetStream("test-image-1.png");
+        var image = TestHelper.GetStream("png image-1.png");
 
         // Act
         shapes.AddPicture(image);
@@ -541,7 +541,7 @@ public class ShapeCollectionTests : SCTest
         // Arrange
         var pres = new Presentation();
         var shapes = pres.Slides[0].Shapes;
-        var image = TestHelper.GetStream("test-image-1.png");
+        var image = TestHelper.GetStream("png image-1.png");
 
         // Act
         shapes.AddPicture(image);
@@ -549,6 +549,22 @@ public class ShapeCollectionTests : SCTest
         // Assert
         var addedPicture = shapes.Last();
         addedPicture.Height.Should().Be(300);
+    }
+
+    [Test]
+    public void AddPicture_adds_picture_with_correct_mime()
+    {
+        // Arrange
+        var pres = new Presentation();
+        var shapes = pres.Slides[0].Shapes;
+        var image = StreamOf("jpeg image.jpg");
+
+        // Act
+        shapes.AddPicture(image);
+
+        // Assert
+        var addedPictureImage = shapes.Last<IPicture>().Image!;
+        addedPictureImage.MIME.Should().Be("image/jpeg");
     }
 
     [Test]