+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.54.2 - Unreleased
 🐞Fixed `Slides.AddEmptySlide()` [#658](https://github.com/ShapeCrawler/ShapeCrawler/issues/658)  
+🐞Fixed `Image.MIME`
 
 ## Version 0.54.1 - 2024-08-28
 🐞Fixed `IFont.LatinName` [#669](https://github.com/ShapeCrawler/ShapeCrawler/issues/669)