+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Version 0.54.2 - Unreleased
 🐞Fixed `Slides.AddEmptySlide()` [#658](https://github.com/ShapeCrawler/ShapeCrawler/issues/658)  
-🐞Fixed `Image.MIME`
+🐞Fixed `Image.MIME`  
+🐞Fixed `Shapes.AddPicture()`
 
 ## Version 0.54.1 - 2024-08-28
 🐞Fixed `IFont.LatinName` [#669](https://github.com/ShapeCrawler/ShapeCrawler/issues/669)