+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Version 0.54.2 - Unreleased
 🐞Fixed `Slides.AddEmptySlide()` [#658](https://github.com/ShapeCrawler/ShapeCrawler/issues/658)  
+🐞Fixed `Shapes.Remove()` [#735](https://github.com/ShapeCrawler/ShapeCrawler/issues/735)  
 🐞Fixed `Image.MIME`  
 🐞Fixed `Shapes.AddPicture()`
+
 
 ## Version 0.54.1 - 2024-08-28
 🐞Fixed `IFont.LatinName` [#669](https://github.com/ShapeCrawler/ShapeCrawler/issues/669)